Added impl for other step types
